f** yeah, it works

+++ lib/dynamic_array/dynamic_array.h
@@ -1,7 +1,7 @@
-#include "../point/point.h"
-
 #ifndef DYNAMIC_ARRAY_H
 #define DYNAMIC_ARRAY_H
+#include "../point/point.h"
+#include <stdlib.h>
 
 #define DYNAMIC_ARRAY_INIT_CAPACITY 4
 
@@ -15,19 +15,19 @@
 
 typedef struct DynamicArray {
     void **items;
-    int capacity;
-    int total;
+    size_t capacity;
+    size_t total;
 } DynamicArray;
 
 int dynamic_array_create(DynamicArray *);
-DynamicArray dynamic_array_slice(DynamicArray * , int );
-int dynamic_array_size(DynamicArray *);
-int dynamic_array_capacity(DynamicArray *);
-static int dynamic_array_resize(DynamicArray *, int);
+DynamicArray dynamic_array_slice(DynamicArray * , size_t );
+size_t dynamic_array_size(DynamicArray *);
+size_t dynamic_array_capacity(DynamicArray *);
+static int dynamic_array_resize(DynamicArray *, size_t);
 int dynamic_array_push(DynamicArray *, void *);
-int dynamic_array_insert(DynamicArray *, int, void *);
-void *dynamic_array_get(DynamicArray *, int);
-int dynamic_array_delete(DynamicArray *, int);
+int dynamic_array_insert(DynamicArray *, size_t, void *);
+void *dynamic_array_get(DynamicArray *, size_t);
+int dynamic_array_delete(DynamicArray *, size_t);
 int dynamic_array_pop(DynamicArray *);
 int dynamic_array_destroy(DynamicArray *);
 //void dynamic_array_print(DynamicArray *);
